backend/requirements.txt
PyMuPDF
openai
python-dotenv
fastapi
uvicorn[standard]
SQLAlchemy
aiosqlite
python-jose[cryptography]
passlib[bcrypt] 